refactor code to auth0 service
